+++ app/common/app.go
@@ -1,44 +1,102 @@
 package common
 
 import (
-	"errors"
+	"context"
 	"fmt"
 	"log/slog"
-	"net/http"
 
-	"github.com/oklog/run"
+	"github.com/google/wire"
 
 	"github.com/openmeterio/openmeter/app/config"
+	"github.com/openmeterio/openmeter/openmeter/app"
+	appadapter "github.com/openmeterio/openmeter/openmeter/app/adapter"
+	appsandbox "github.com/openmeterio/openmeter/openmeter/app/sandbox"
+	appservice "github.com/openmeterio/openmeter/openmeter/app/service"
+	appstripe "github.com/openmeterio/openmeter/openmeter/app/stripe"
+	appstripeadapter "github.com/openmeterio/openmeter/openmeter/app/stripe/adapter"
+	appstripeservice "github.com/openmeterio/openmeter/openmeter/app/stripe/service"
+	"github.com/openmeterio/openmeter/openmeter/customer"
+	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
+	"github.com/openmeterio/openmeter/openmeter/namespace"
+	"github.com/openmeterio/openmeter/openmeter/secret"
 )
 
-// Metadata provides information about the service to components that need it (eg. telemetry).
-type Metadata struct {
-	ServiceName       string
-	Version           string
-	Environment       string
-	OpenTelemetryName string
-}
+var App = wire.NewSet(
+	wire.FieldsOf(new(config.Configuration), "Apps"),
 
-func NewMetadata(conf config.Configuration, version string, serviceName string) Metadata {
-	return Metadata{
-		ServiceName:       fmt.Sprintf("openmeter-%s", serviceName),
-		Version:           version,
-		Environment:       conf.Environment,
-		OpenTelemetryName: fmt.Sprintf("openmeter.io/%s", serviceName),
+	NewAppService,
+	NewAppStripeService,
+	NewAppSandbox,
+)
+
+func NewAppService(logger *slog.Logger, db *entdb.Client, appsConfig config.AppsConfiguration) (app.Service, error) {
+	// TODO: remove this check after enabled by default
+	if !appsConfig.Enabled || db == nil {
+		return nil, nil
 	}
-}
 
-// Runner is a helper struct that runs a group of services.
-type Runner struct {
-	Group  run.Group
-	Logger *slog.Logger
-}
-
-func (r Runner) Run() {
-	err := r.Group.Run()
-	if e := (run.SignalError{}); errors.As(err, &e) {
-		r.Logger.Info("received signal: shutting down", slog.String("signal", e.Signal.String()))
-	} else if !errors.Is(err, http.ErrServerClosed) {
-		r.Logger.Error("application stopped due to error", slog.String("error", err.Error()))
+	appAdapter, err := appadapter.New(appadapter.Config{
+		Client:  db,
+		BaseURL: appsConfig.BaseURL,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to create app adapter: %w", err)
 	}
+
+	return appservice.New(appservice.Config{
+		Adapter: appAdapter,
+	})
+}
+
+func NewAppStripeService(logger *slog.Logger, db *entdb.Client, appsConfig config.AppsConfiguration, appService app.Service, customerService customer.Service, secretService secret.Service) (appstripe.Service, error) {
+	// TODO: remove this check after enabled by default
+	if !appsConfig.Enabled || db == nil {
+		return nil, nil
+	}
+
+	appStripeAdapter, err := appstripeadapter.New(appstripeadapter.Config{
+		Client:          db,
+		AppService:      appService,
+		CustomerService: customerService,
+		SecretService:   secretService,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to create appstripe adapter: %w", err)
+	}
+
+	return appstripeservice.New(appstripeservice.Config{
+		Adapter:       appStripeAdapter,
+		AppService:    appService,
+		SecretService: secretService,
+	})
+}
+
+func NewAppSandbox(ctx context.Context, logger *slog.Logger, appsConfig config.AppsConfiguration, appService app.Service, namespaceManager *namespace.Manager) (*appsandbox.App, error) {
+	if !appsConfig.Enabled {
+		return nil, nil
+	}
+
+	_, err := appsandbox.NewFactory(appsandbox.Config{
+		AppService: appService,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize app sandbox factory: %w", err)
+	}
+
+	app, err := appsandbox.AutoProvision(ctx, appsandbox.AutoProvisionInput{
+		Namespace:  namespaceManager.GetDefaultNamespace(),
+		AppService: appService,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to auto-provision sandbox app: %w", err)
+	}
+
+	logger.Info("sandbox app auto-provisioned", "app_id", app.GetID().ID)
+
+	appSandbox, ok := app.(appsandbox.App)
+	if !ok {
+		return nil, fmt.Errorf("failed to cast app to sandbox app")
+	}
+
+	return &appSandbox, nil
 }
