+++ app/common/customer.go
@@ -4,13 +4,24 @@ import (
 	"fmt"
 	"log/slog"
 
+	"github.com/google/wire"
+
 	"github.com/openmeterio/openmeter/openmeter/customer"
 	customeradapter "github.com/openmeterio/openmeter/openmeter/customer/adapter"
 	customerservice "github.com/openmeterio/openmeter/openmeter/customer/service"
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
 )
 
+var Customer = wire.NewSet(
+	NewCustomerService,
+)
+
 func NewCustomerService(logger *slog.Logger, db *entdb.Client) (customer.Service, error) {
+	// TODO: remove this check after enabled by default
+	if db == nil {
+		return nil, nil
+	}
+
 	customerAdapter, err := customeradapter.New(customeradapter.Config{
 		Client: db,
 		Logger: logger.WithGroup("customer.postgres"),
