+++ app/common/meter.go
@@ -1,0 +1,13 @@
+package common
+
+import (
+	"github.com/samber/lo"
+
+	"github.com/openmeterio/openmeter/openmeter/meter"
+	"github.com/openmeterio/openmeter/pkg/models"
+	"github.com/openmeterio/openmeter/pkg/slicesx"
+)
+
+func NewMeterRepository(meters []*models.Meter) *meter.InMemoryRepository {
+	return meter.NewInMemoryRepository(slicesx.Map(meters, lo.FromPtr[models.Meter]))
+}
