+++ app/common/namespace.go
@@ -1,0 +1,24 @@
+package common
+
+import (
+	"fmt"
+
+	"github.com/openmeterio/openmeter/app/config"
+	"github.com/openmeterio/openmeter/openmeter/namespace"
+)
+
+func NewNamespaceManager(
+	handlers []namespace.Handler,
+	conf config.NamespaceConfiguration,
+) (*namespace.Manager, error) {
+	manager, err := namespace.NewManager(namespace.ManagerConfig{
+		Handlers:          handlers,
+		DefaultNamespace:  conf.Default,
+		DisableManagement: conf.DisableManagement,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("create namespace manager: %v", err)
+	}
+
+	return manager, nil
+}
