+++ app/common/secret.go
@@ -3,14 +3,19 @@ package common
 import (
 	"log/slog"
 
-	"github.com/openmeterio/openmeter/app/config"
+	"github.com/google/wire"
+
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
 	"github.com/openmeterio/openmeter/openmeter/secret"
 	secretadapter "github.com/openmeterio/openmeter/openmeter/secret/adapter"
 	secretservice "github.com/openmeterio/openmeter/openmeter/secret/service"
 )
 
-func NewSecretService(logger *slog.Logger, db *entdb.Client, appsConfig config.AppsConfiguration) (secret.Service, error) {
+var Secret = wire.NewSet(
+	NewSecretService,
+)
+
+func NewSecretService(logger *slog.Logger, db *entdb.Client) (secret.Service, error) {
 	secretAdapter := secretadapter.New()
 
 	return secretservice.New(secretservice.Config{
