+++ app/common/wire.go
@@ -22,7 +22,6 @@ import (
 )
 
 var Config = wire.NewSet(
-	wire.FieldsOf(new(config.Configuration), "Apps"),
 	wire.FieldsOf(new(config.Configuration), "Aggregation"),
 	wire.FieldsOf(new(config.AggregationConfiguration), "ClickHouse"),
