+++ app/config/config.go
@@ -160,7 +160,6 @@ func SetViperDefaults(v *viper.Viper, flags *pflag.FlagSet) {
 	ConfigurePortal(v)
 	ConfigureEvents(v)
 	ConfigureBalanceWorker(v)
-	ConfigureBillingWorker(v)
 	ConfigureNotification(v)
 	ConfigureBilling(v)
 	ConfigureProductCatalog(v)
