+++ cmd/billing-worker/.air.toml
@@ -4,8 +4,8 @@ tmp_dir = "tmp"
 
 [build]
   args_bin = ["--config", "./config.yaml", "--telemetry-address", ":10002"]
-  bin = "./tmp/openmeter-balance-worker"
-  cmd = "go build -tags dynamic -o ./tmp/openmeter-balance-worker ./cmd/balance-worker"
+  bin = "./tmp/openmeter-billing-worker"
+  cmd = "go build -tags dynamic -o ./tmp/openmeter-billing-worker ./cmd/billing-worker"
   delay = 0
   exclude_dir = ["assets", "ci", "deploy", "docs", "examples", "testdata", "quickstart", "tmp", "vendor", "api/client", "node_modules"]
   exclude_file = []
